+++ app/controllers/staff/report_forecasts_controller.rb
@@ -7,22 +7,16 @@ class Staff::ReportForecastsController < Staff::BaseController
     @report = Report.find(params["report_id"])
     authorize @report
 
+    forecasts = @report.forecasts_for_reportable_activities
+
     @report_presenter = ReportPresenter.new(@report)
     @report_activities = @report.reportable_activities
-    @total_forecast = TotalPresenter.new(forecasts.sum(&:value)).value
+    @total_forecast = @report_presenter.summed_forecasts_for_reportable_activities
     @grouped_forecasts = forecasts
       .includes([:parent_activity])
       .map { |forecast| ForecastPresenter.new(forecast) }
       .group_by { |forecast| ActivityPresenter.new(forecast.parent_activity) }
 
     render "staff/reports/forecasts"
-  end
-
-  private
-
-  def forecasts
-    ForecastOverview
-      .new(@report_activities.map(&:id))
-      .latest_values
   end
 end