+++ db/migrate/20210630094140_add_report_id_to_historical_event.rb
@@ -1,0 +1,5 @@
+class AddReportIdToHistoricalEvent < ActiveRecord::Migration[6.1]
+  def change
+    add_reference :historical_events, :report, type: :uuid
+  end
+end