+++ CHANGELOG.md
@@ -714,9 +714,25 @@
 - Change GCRF Strategic Area codes to Alphanumeric
 - Record edits to Activities made via Wizard form, using new HistoricalEvent entity
 
+## [release-60] - 2021-07-06
+
+- Show actuals grouped by activity against a specific report
+- Show uploaded transactions after an upload
+- Record edits to Activities made in bulk via CSV imports, using the new Historical Event entity
+- Associate Report with HistoricalEvent when editing Activity via Wizard
+- BEIS users have a home page that lets them view activities by delivery partner
+  organisation
+- Make sure a Budget has a Budget Type
+- Delivery partner users have a home page that lets them view and search their
+  own activities
+- Associate Report with HistoricalEvent when uploading a bulk CSV
+- Add a report summary tab
+- List an Activity's historical events in a new "Change history" tab 
+
 ## [unreleased]
 
-[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-59...HEAD
+[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-60...HEAD
+[release-60]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-59...release-60
 [release-59]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-58...release-59
 [release-58]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-57...release-58
 [release-57]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-56...release-57